Second Submission for the assignment
Got a lot of points off due to the wrong answer when i realized that most of these questions were only asking for analysis of the top 15 which referenced the dataframe created after question 1! after i changed the functions that generated my data frames I barely had to make any changes to get the correct answers. Formatting of question 12 and 13 are still incorrect in some way - need to work on that next.